% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatec.R
\name{ar.table}
\alias{ar.table}
\title{Construct a table with models for the given assessment report}
\usage{
ar.table(assessment = 2)
}
\description{
Construct a table with models for the given assessment report
}
