% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatec.R
\name{ar.text}
\alias{ar.text}
\title{Web link for Assesment report  (AR 1 to AR 4)}
\usage{
ar.text(assessment = 2)
}
\arguments{
\item{assessment}{A numeric value associated to the assessment report}
}
\value{
return the web link to all the model associated to this assessment report
}
\description{
Prepare the web link to call for the models associated to each assessment report.
}
