% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatec.R
\name{AR1_4.get.all}
\alias{AR1_4.get.all}
\title{Get ALL the information from EC server based Latitude and Longitude.}
\usage{
AR1_4.get.all(Latitude, Longitude)
}
\description{
Get ALL the information from EC server based Latitude and Longitude.
}
