% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatec.R
\name{AR1_4.get}
\alias{AR1_4.get}
\title{Get the information from EC server based Latitude and Longitude for one index row.}
\usage{
AR1_4.get(Latitude, Longitude, i = 1)
}
\description{
Get the information from EC server based Latitude and Longitude for one index row.
}
