% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatec.R
\name{AR1_4.get.results}
\alias{AR1_4.get.results}
\title{Process the results from the files from AR 1 to AR 4}
\usage{
AR1_4.get.results()
}
\description{
Process the results from the files from AR 1 to AR 4
}
