% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatec.R
\name{AR1_4.Index}
\alias{AR1_4.Index}
\title{Index table for AR 1 to AR 4}
\format{Index table for AR1 to AR 4}
\usage{
data(AR1_4.Index)
}
\description{
Index table for AR 1 to AR 4
}
\references{
EC
}
\author{
Victor Munoz
}
