% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatec.R
\name{AR1_4.year.setup}
\alias{AR1_4.year.setup}
\title{Year setup for AR 1 to 4.}
\usage{
AR1_4.year.setup(Baseline.start = 1960, Baseline.end = 1990,
  Analysis.start = 2041, Analysis.end = 2070)
}
\description{
Year setup for AR 1 to 4.
}
