% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatec.R
\name{AR5.get.results}
\alias{AR5.get.results}
\title{Report all the results for the netcdf files based on a index file and parameters}
\usage{
AR5.get.results(path, Baseline.start, Baseline.end, Analysis.start,
  Analysis.end, Latitude, Longitude, Parameter)
}
\description{
Report all the results for the netcdf files based on a index file and parameters
}
