% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliar Tools.R
\name{BING.elev}
\alias{BING.elev}
\title{Distance from the elevation based on Bing from a lat and long coordinate}
\usage{
BING.elev(longitude = 30, latitude = 40)
}
\arguments{
\item{longitude}{Longitude of the point}

\item{latitude}{Latitude of the point}
}
\description{
Calculate the elevation based on Bing
}
