% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatec.R
\name{cc.text}
\alias{cc.text}
\title{Function which read the information from EC based on defined variables}
\usage{
cc.text(assessment = 2, model.1 = 23, model.2 = 1, experiment = 16,
  variable = 11, basef = 1961, baset = 1990, lat = 52.7,
  long = -95.25)
}
\description{
Function which read the information from EC based on defined variables
}
