% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliar Tools.R
\name{coast.dist}
\alias{coast.dist}
\title{Distance from the coast from a lat and long coordinate}
\usage{
coast.dist(longitude = 30, latitude = -40)
}
\arguments{
\item{longitude}{Longitude of the point}

\item{latitude}{Latitude of the point}
}
\description{
Calculate the distance from the coast
}
