% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliar Tools.R
\name{daily2monthly}
\alias{daily2monthly}
\title{Transform a daily zoo in a monthly zoo}
\usage{
daily2monthly(x = flow.daily, FUN = sum, na.rm = TRUE, max.miss = 5)
}
\arguments{
\item{FUN}{Funtion to operation (mean, sum)}

\item{na.rm}{Uses of NA}

\item{max.miss}{Maximum amount of missing days in time series to consider a month}

\item{X}{Daily time series in zoo format}
}
\value{
data frame with outputs
}
\description{
Transform a daily zoo in a monthly zoo
}
