% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECmeteo.R
\name{degtocompass}
\alias{degtocompass}
\title{Transform degrees in compass direction}
\usage{
degtocompass(num)
}
\arguments{
\item{num}{Degree value of wind speed to be transformed in compass direction}
}
\description{
Transform degrees in compass direction
}
