% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECflow.R
\name{ECf.add.patched}
\alias{ECf.add.patched}
\title{Corrected the record adding patched information to ECflow.db}
\usage{
ECf.add.patched(flow, idx)
}
\arguments{
\item{flow}{flow from the patched station (ECf.patch$cut.ts) or excel}

\item{idx}{row to be replaced with respect to ECflow.idx}
}
\description{
Corrected the record adding patched information to ECflow.db
}
