% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECflow.R
\name{ECf.add}
\alias{ECf.add}
\title{Add Gauge Stations (EC)}
\usage{
ECf.add(x = stations.id, monthly.zero = NA)
}
\arguments{
\item{stations.id}{Station web id from EC}
}
\description{
Add one specific station to the work database
}
