% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliar Tools.R
\name{ECf.coast.dist}
\alias{ECf.coast.dist}
\title{Distance from the coast from a lat and long coordinate for the ECflow.idx}
\usage{
ECf.coast.dist()
}
\description{
Calculate the distance from the coast for ECflow.idx
}
