% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECflow.R
\name{ECf.find}
\alias{ECf.find}
\title{Find Gauge Stations (EC)}
\usage{
ECf.find(lat = 49.6, lon = -114.5, n = 3)
}
\arguments{
\item{lat}{Latitude in decimal degrees}

\item{n}{Amount of station to be reviewed}

\item{long}{Longitude in decimal degrees (in Canada is negative)}
}
\description{
Find the closest stations n from the HYDAT database based on latitude and longitude
}
