% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliar Tools.R
\name{ECf.GE.elev}
\alias{ECf.GE.elev}
\title{Estimate the ground elevation base on Google Earth as a input for ECflow.idx}
\usage{
ECf.GE.elev()
}
\description{
Estimate the ground elevation base on Google Earth as a input for ECflow.idx
}
