% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Low flow.r
\name{ECf.low}
\alias{ECf.low}
\title{Calculate low flow for all the gauges at ECflow.db}
\usage{
ECf.low(n = 7)
}
\arguments{
\item{n}{number of days for minimum stats.}
}
\description{
Calculate low flow for all the gauges at ECflow.db
}
