% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Low flow.r
\name{ECf.low.ts}
\alias{ECf.low.ts}
\title{Calculate low flow for all a time series}
\usage{
ECf.low.ts(ts, n = 7)
}
\arguments{
\item{n}{number of days for minimum stats.}
}
\description{
Calculate low flow for all a time series
}
