% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECflow.R
\name{ECf.out}
\alias{ECf.out}
\title{Export the results from analysis}
\usage{
ECf.out(filename = "output.xlsx", FA.results = FALSE, Wat.results = FALSE)
}
\arguments{
\item{filename}{Excel filename to export the results}
}
\description{
Export the results from analysis
}
