% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECflow.R
\name{ECf.patch}
\alias{ECf.patch}
\title{Patch information between gauge stations}
\usage{
ECf.patch(to.be.patched = 2, with = 4, poly.n = 0.75)
}
\arguments{
\item{to.be.patched}{= row index from ECflow.idx of the station to be patched}

\item{with}{= row index from ECflow.idx of the station to be patched with}

\item{poly.n}{= polynomial degree for the fitting curve}
}
\description{
Patch information between gauge stations
}
