% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECflow.R
\name{ECf.rm}
\alias{ECf.rm}
\title{Remove the information from the work database}
\usage{
ECf.rm(x = NULL)
}
\arguments{
\item{x}{Vector with the amount to be removed}
}
\description{
Remove the information from the work database
}
