% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECflow.R
\name{ECf.select.min}
\alias{ECf.select.min}
\title{Select stations which have n amount minimum amount of info}
\usage{
ECf.select.min(n = 0)
}
\arguments{
\item{n}{Minimun amount of years}
}
\description{
Select stations which have n amount minimum amount of info
}
