% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECmeteo.R
\name{ECm.analysis}
\alias{ECm.analysis}
\title{Meteorological Data Analysis}
\usage{
ECm.analysis<-function(db, FUN.month, Min.days.annual, Max.miss.days.month, output, EC, matrix.start.year)
}
\arguments{
\item{db}{the database from \code{\link[Hydro]{ECm.prepare}} that contains the information for the selected meteorological parameter}

\item{FUN.month}{function used to convert daily data into monthly and annual data. Options: sum, mean}

\item{Min.days.annual}{the minimum number of days in a year for the data to be included in the analysis. Suggested: 200-365}

\item{output}{the name of the returned database}

\item{EC}{specify if the original data was sourced from Environment Canada. If TRUE, function uses information from "selected.m.network" database}

\item{matrix.start.year}{the start year of the graph presenting available information. Does not affect tabular output.}

\item{Max.days.month}{the maximum number of missing days in a month for the data to be included in the analysis. Suggested: 0-5}
}
\value{
Three returns:
1. Graph summarizing years with data for every station. \cr
2. Database of compiled meteorological information (see details below). \cr
3. Summary of the stations analyzed with annual information. \cr

Database Elements:
\itemize{
\item Daily.Comp - daily zoo object of available meteorological information.
\item Monthly.avg - monthly averages of available meteorological information.
\item Monthly.avg.nd - monthly averages presented as a percentage of the year (i.e. non dimensional).
\item Monthly.table - monthly information for every year and station.
\item Month.Comp - monthly zoo object of available meteorological information.
\item MWI - summary of the months with information for each station, supported with average and max/min.
\item Annual - annual information for every station.
\item Annual.comp - annual zoo object of available meteorological information.
\item Annual.Syn - similar table to function output. Describes \code{\link[nsRFA]{Lmoments}} and identifies the month of the min and max, and the rate (i.e. max/min) to describe the shape of hydrograph.
\item Max.Daily.data - presents the daily maximum per year for each station.
\item Max.Daily.table - presents the daily maximum per year for each station as a table.
}
Please note that data from incomplete years and months has been omitted in each of the database elements.
}
\description{
This function compiles meteorological information on different timescales for the selected parameter and presents data in graphical and tabular form.
Information is sourced from the database that was previously prepared by \code{\link[Hydro]{ECm.prepare}} or edited by \code{\link[Hydro]{ECm.rm}}. Data is processed on daily, monthly and annual timescales.
Daily information may be converted to monthly and annual data by sumation or averaging. Data is omitted if the minimum days in a year or month criteria is not fulfilled.
}
\examples{
## EX 1: Analyze the precipitation totals for EC stations. Consider a full year of data as 360 days, and a maximum of 5 missing days per month. Present the graph starting in the year 1980.
ECm.analysis(db=Total.Precip..mm.db, FUN.month=sum, Min.days.annual=360, Max.miss.days.month=5, output="station.syn", EC=TRUE, matrix.start.year=1980)

}
