% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECmeteo.R
\name{ECm.capture.all}
\alias{ECm.capture.all}
\title{Download Station Information}
\usage{
ECm.capture.all(multicore, timeframe)
}
\arguments{
\item{multicore}{True or false; State whether the function may utilize multicore capabilities for faster processing.}

\item{timeframe}{1: Hourly, 2: Daily, 3: Monthly}
}
\value{
a dataframe consisting of all available meteorological data for each station.
}
\description{
Function to capture the available monthly, daily or hourly meteorological information from the selected meteorological stations using \code{\link[Hydro]{ECm.find}}. Uses \code{\link[Hydro]{ECm.capture}} with the option to use multiple processors.
}
\examples{
## EX 1: Return all hourly data for previous selected 'n' stations.
Assume multicore capabilities.
ECm.capture.all(multicore = T, timeframe = 1)
}
