% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECmeteo.R
\name{ECm.evap}
\alias{ECm.evap}
\title{Estimate Evaporation from a EC information}
\usage{
ECm.evap(stationid = 1174, station.normal = ECm.normal(stationid =
  1174)$table)
}
\arguments{
\item{stationid}{Latitude in decimal degrees}

\item{station.normal}{Normal table from ECn.normal()}
}
\description{
Estimate Evaporation from a EC information
}
