% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECmeteo.R
\name{ECm.find}
\alias{ECm.find}
\title{Find Meteorological Stations from Environment Canada}
\usage{
ECm.find(lat, lon, n, timeframe)
}
\arguments{
\item{lat}{Latitude in decimal degrees.}

\item{lon}{Longitude in decimal degrees. Note: Canadian longitudes are negative.}

\item{n}{The number of stations to be located.}

\item{timeframe}{1: Hourly, 2: Daily}
}
\value{
a dataframe of 'n' stations
}
\description{
Function for locating the nearest 'n' stations to a specified coordinate. Returns detailed list of Canadian stations that have available meteorological data in hourly or daily format.
}
\examples{
## EX. 1: Find the closest 20 stations to 49.6, -114.5 that have hourly data.
ECm.find(lat=49.6, lon=-114.5, n=20, timeframe=1)

## EX. 2: Find the closest 10 stations to Ottawa, Canada that have daily data.
ECm.find(lat=45.42, lon = -75.70, n=10, timeframe=2)
}
