% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECmeteo.R
\name{ECm.id}
\alias{ECm.id}
\title{Station Description}
\usage{
ECm.id(stationid = 1166)
}
\arguments{
\item{stationid}{Station ID Number}
}
\description{
This function summarizes the information about one meteorological station. Provides coordinates, elevation, and identifiers.
}
