% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECmeteo.R
\name{ECm.normal.capture}
\alias{ECm.normal.capture}
\title{Estimate Evaporation from a EC information}
\usage{
ECm.normal.capture(lat = 49.645318, lon = -114.453425, n = 5)
}
\arguments{
\item{lat}{Latitude in decimal degrees}

\item{lon}{Longitude in decimal degrees}

\item{n}{Amount of station to be review}
}
\description{
Estimate Evaporation from a EC information
}
