% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECmeteo.R
\name{ECm.normal}
\alias{ECm.normal}
\title{Climate Normals (One Station)}
\usage{
ECm.normal(stationid = 1174, raw = FALSE, year.cn = 1981)
}
\arguments{
\item{stationid}{Station ID Number}

\item{year.cn}{Start year for the different periods of the climate normals (1981-2010, 1971-2000, 1961-1990). Options: 1981, 1971, 1961}
}
\value{
a table of climate normals for the given station and period.
}
\description{
This function obtains the climate normals for one meteorological station. Sourced from Environment Canada.
}
