% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECmeteo.R
\name{ECm.out}
\alias{ECm.out}
\title{Export the results from analysis}
\usage{
ECm.out(db = Total.Precip..mm.syn, filename = "output.xlsx", FA = F,
  EC = T)
}
\arguments{
\item{filename}{Excel filename to export the results}
}
\description{
Export the results from analysis
}
