% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Frequency Analysis.R
\name{ECm.pf}
\alias{ECm.pf}
\title{Frequency analysis for meteorology at the complete work database}
\usage{
ECm.pf(db = ECm.annual.peak.table, Dailyto24 = 1.13)
}
\arguments{
\item{Dailyto24}{Daily to 24 hours correction default of 1.13}

\item{a}{ECm.annual.peak.table}
}
\description{
Frequency analysis results for the complete meteorologicaldatabase of peak flows
}
