% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECmeteo.R
\name{ECm.prepare}
\alias{ECm.prepare}
\title{Prepare for Analysis}
\usage{
ECm.prepare(db, Parameter.name)
}
\arguments{
\item{db}{The database of daily or hourly meteorological information obtained from \code{\link[Hydro]{ECm.capture.all}}.}

\item{Parameter.name}{The selected meteorological parameter name from \code{\link[Hydro]{ECm.variables}}.}
}
\description{
This function prepares the database for use in \code{\link[Hydro]{ECm.analysis}} based on the selected meteorological parameter by removing all other parameters from the database.
}
\examples{
## EX 1: Prepare database for Mean Temperature and remove all other parameters
ECm.prepare(db=ECm.stations.daily, Parameter.name="Mean.Temp..Â.C.")
}
