% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECmeteo.R
\name{ECm.prepare.subdaily}
\alias{ECm.prepare.subdaily}
\title{Sub-Daily Preparation}
\usage{
ECm.prepare.subdaily()
}
\value{
a database of daily information
}
\description{
This function converts the database of sub-daily information to daily information for use in \code{\link[Hydro]{ECm.analysis}}. Parameters requiring sub-daily processing include Wind Speed and Wind Direction.
Wind speed and direction are averaged appropriately. Other parameters available sub-daily are averaged using mean.
}
