% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECmeteo.R
\name{ECm.rm}
\alias{ECm.rm}
\title{Remove Information}
\usage{
ECm.rm(db, x, EC)
}
\arguments{
\item{db}{The database of meteorological information from \code{\link[Hydro]{ECm.prepare}}}

\item{x}{Criteria for the information to be removed. Reference $MWI from \code{\link[Hydro]{ECm.analysis}}}

\item{EC}{Specify if the original data was sourced from Environment Canada. If TRUE, function edits information from "selected.m.network" database}
}
\description{
This function removes stations from the database returned from \code{\link[Hydro]{ECm.prepare}} based on the number of months of information. Used to determine useful data.\cr
Please Note: \code{\link[Hydro]{ECm.analysis}} should be run before and after this function.
}
\examples{
## EX 1: Remove stations that have an average months with information less than 10. Using precipitation totals sourced from EC.
ECm.rm(db="Total.Precip..mm.db", Total.Precip..mm.syn$MWI$MWI.Avg<10, EC=TRUE)
}
