% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECmeteo.R
\name{ECm.select.min}
\alias{ECm.select.min}
\title{Select stations which have n amount minimum amount of info}
\usage{
ECm.select.min(n = 0)
}
\arguments{
\item{n}{Minimun amount of years}
}
\description{
Select stations which have n amount minimum amount of info
}
