% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECmeteo.R
\name{ECm.text.n}
\alias{ECm.text.n}
\title{Source URL for EC Climate Normals (One Station)}
\usage{
ECm.text.n(stationid = 1174, year.cn = 1981)
}
\arguments{
\item{stationid}{Station ID Number}

\item{year.cn}{Start year for the different periods of the climate normals (1981-2010, 1971-2000, 1961-1990). Options: 1981, 1971, 1961}
}
\description{
This function provides the URL text for the climate normals from a single meteorological station, sourced from Environment Canada.
}
