% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECmeteo.R
\name{ECm.text}
\alias{ECm.text}
\title{Source URL for EC Meteorological Data (One Station)}
\usage{
ECm.text(stationid = 1166, year = 1956, timeframe = 1)
}
\arguments{
\item{stationid}{Station ID Number}

\item{year}{Selected year of data}

\item{timeframe}{1: Hourly, 2: Daily, 3: Monthly}
}
\description{
This function provides the URL text for the historical data from a single meteorological station, sourced from Environment Canada.
}
