% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECmeteo.R
\name{ECm.variables}
\alias{ECm.variables}
\title{Meteorological Variable Names}
\usage{
ECm.variables(db)
}
\arguments{
\item{db}{The database of station meteorological information returned from \code{\link[Hydro]{ECm.capture.all}}}
}
\value{
a character vector of meteorological variables
}
\description{
Displays the meteorological variables that are available at the stations that can be used for input into \code{\link[Hydro]{ECm.prepare}}.
}
