% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reanalisys.R
\name{ERA.Interim.2014}
\alias{ERA.Interim.2014}
\title{Index table for Reanalysis ERA-Interim 2014}
\format{Index table for Reanalysis ERA-Interim 2014}
\usage{
data(ERA.Interim.2014)
}
\description{
Index table for Reanalysis ERA-Interim 2014
}
\references{
ECMWF
}
\author{
Victor Munoz
}
