% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evap Corrections.R
\name{ET.MortonCRAE}
\alias{ET.MortonCRAE}
\title{ET.MortonCRAE for High latitudes}
\usage{
ET.MortonCRAE(data, constants, ts = "monthly", est = "potential ET",
  solar = "sunshine hours", Tdew = TRUE, alpha = NULL, ...)
}
\description{
ET.MortonCRAE for High latitudes
}
