% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evap Corrections.R
\name{ET.MortonCRWE}
\alias{ET.MortonCRWE}
\title{ET.MortonCRWE for High latitudes}
\usage{
ET.MortonCRWE(data, constants, ts = "monthly", est = "potential ET",
  solar = "sunshine hours", Tdew = TRUE, alpha = NULL, ...)
}
\description{
ET.MortonCRWE for High latitudes
}
