% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evap Corrections.R
\name{ET.PenmanMonteith}
\alias{ET.PenmanMonteith}
\title{ET.PenmanMonteith for High latitudes}
\usage{
ET.PenmanMonteith(data, constants, ts = "daily", solar = "sunshine hours",
  wind = "yes", crop = "short", ...)
}
\description{
ET.PenmanMonteith for High latitudes
}
