% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro.R
\name{ETSR}
\alias{ETSR}
\title{Extraterrestrial solar radiation, per Shuttleworth 1993}
\usage{
ETSR(psi, J)
}
\arguments{
\item{psi}{Latitude, in degrees}

\item{J}{Julian day}
}
\value{
Solar radiation in mm/day
}
\description{
Extraterrestrial solar radiation, per Shuttleworth 1993
}
