% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatec.R
\name{ex.text}
\alias{ex.text}
\title{Web link for Experiment (AR 1 to AR 4)}
\usage{
ex.text(row = 1)
}
\arguments{
\item{row}{Row value for the Index table associated to AR1 to AR4}
}
\value{
return the web link to the experiment associated to the DB row
}
\description{
Prepare the web link to call for the experiment associated to each model and assessment report.
}
