% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Frequency Analysis.R
\name{FA.clip}
\alias{FA.clip}
\title{Frequency Analysis from Clipboard  information by columns}
\usage{
FA.clip()
}
\arguments{
\item{information}{from clipboard}
}
\value{
Frequency analysis results at the clipboard
}
\description{
Estimate the Frequency Analysis from Clipboard information by columns
}
