% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Frequency Analysis.R
\name{FA.ECflow.1}
\alias{FA.ECflow.1}
\title{Frequency analysis for flows at one station}
\usage{
FA.ECflow.1(stationid)
}
\arguments{
\item{stationid}{Station id to prepare the frequency analysis required a flow analysis results}
}
\description{
Estimate frequency analysis for peak flows at one station
}
