% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Frequency Analysis.R
\name{FA.ECflow.all}
\alias{FA.ECflow.all}
\title{Frequency analysis for flows at the complete work database}
\usage{
FA.ECflow.all()
}
\description{
Flow Analysis provide results of a complete database of peak flows
}
