% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Frequency Analysis.R
\name{FA.ECflow.graph}
\alias{FA.ECflow.graph}
\title{Frequency analysis graph for results at flow.freq.graph}
\usage{
FA.ECflow.graph(event = "Dry_200y", max.distance = 400)
}
\arguments{
\item{event}{string of the event based on the titles from table flow.freq.graph}

\item{max.distance}{number in km for the maximum distance considered for the graph}
}
\description{
Frequency analysis graph for results at flow.freq.graph
}
