% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Frequency Analysis.R
\name{FA}
\alias{FA}
\title{Frequency Analysis}
\usage{
FA(a, distrib = c("NORM", "LN", "GEV", "GUMBEL", "P3", "LP3"),
  display = FALSE)
}
\arguments{
\item{a}{numerical vector}

\item{display}{Option for graphical output (T/F)}
}
\description{
Estimate the Frequency Analysis
}
