% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Watershed tools.R
\name{gdal_polygonizeR}
\alias{gdal_polygonizeR}
\title{Transform from a raste to a vector with gdal}
\usage{
gdal_polygonizeR(x, outshape = NULL, gdalformat = "ESRI Shapefile",
  pypath = NULL, readpoly = TRUE, quiet = TRUE)
}
\description{
Transform from a raste to a vector with gdal
}
