% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliar Tools.R
\name{IDF}
\alias{IDF}
\title{IDF function – Maximum annual for a define time step
IDF function – Maximum annual for a define time step}
\usage{
IDF(Pp, n)
}
\description{
IDF function – Maximum annual for a define time step
IDF function – Maximum annual for a define time step
}
