% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mann-kendall analysis.R
\name{MannKendall.eval}
\alias{MannKendall.eval}
\title{Mann-Kendall Evaluation for Climate Change analysis}
\usage{
MannKendall.eval(ts, prob = 0.95, Baseline.start, Baseline.end,
  Analysis.start, Analysis.end)
}
\arguments{
\item{ts}{Zoo time series}

\item{prob}{Probability to evalution}
}
\value{
return a table with the parameter change
}
\description{
Prepare a Mann-Kendall analysis based on time series.
}
