% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mann-kendall analysis.R
\name{MannKendall.std}
\alias{MannKendall.std}
\title{Mann-Kendall Results from Time series}
\usage{
MannKendall.std(ts, prob = 0.95)
}
\arguments{
\item{ts}{Zoo time series}

\item{prob}{Probability to evalution}
}
\value{
return the web link to all the model associated to this assessment report
}
\description{
Prepare a Mann-Kendall analysis based on time series.
}
