% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evap Corrections.R
\name{MERRA.Evap.HL}
\alias{MERRA.Evap.HL}
\title{MERRA evaporation for High latitudes}
\usage{
MERRA.Evap.HL(Meteo.info = MERRA.table, MAP = NULL, Latitude = 39.346,
  Longitude = 34)
}
\description{
MERRA evaporation for High latitudes
}
