% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MERRA evaporation.R
\name{MERRA.get}
\alias{MERRA.get}
\title{Get meteorological information from MERRA (NASA POWER)}
\usage{
MERRA.get(lat = 33.419, long = 110.974164)
}
\description{
Get meteorological information from MERRA (NASA POWER)
}
