% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MERRA evaporation.R
\name{MERRA.process}
\alias{MERRA.process}
\title{Process meteorological information from MERRA (NASA POWER)}
\usage{
MERRA.process(x = MERRA.file)
}
\description{
Process meteorological information from MERRA (NASA POWER)
}
