% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MERRA evaporation.R
\name{MERRA.text}
\alias{MERRA.text}
\title{Find meteorological information from MERRA (NASA POWER)}
\usage{
MERRA.text(lat = 33.419, long = 110.974164)
}
\description{
Find meteorological information from MERRA (NASA POWER)
}
