% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliar Tools.R
\name{Meteo.patch}
\alias{Meteo.patch}
\title{Patch hydrological time series
Patch hydrological time series}
\usage{
Meteo.patch(df = flow.all.zoo, areas, to.be.patched = 2, with = 4,
  poly.n = 0.75, zero.interception = FALSE, lower.than.zero.is.zero = T,
  monthly.FUN = mean, units = "m3/s", setup = 1)
}
\description{
Patch hydrological time series
Patch hydrological time series
}
