% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliar Tools.R
\name{monthly2daily}
\alias{monthly2daily}
\title{Divide the information from monthly to daily}
\usage{
monthly2daily(timeseries)
}
\arguments{
\item{timeseries}{Zoo time series}
}
\description{
Divide the information from monthly to daily
}
