% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECflow.R
\name{NAf.add}
\alias{NAf.add}
\title{Add Gauge Stations (EC&USGS)}
\usage{
NAf.add(x = stations.id, monthly.zero = NA)
}
\arguments{
\item{stations.id}{Station web id from EC & USGS}
}
\description{
Add one specific station to the work database
}
