% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECflow.R
\name{NAf.find}
\alias{NAf.find}
\title{Find Gauge Stations in Northamerica Environment Canada & USGS}
\usage{
NAf.find(lat = 62.895278, lon = -139.377426, n = 100)
}
\arguments{
\item{lat}{Latitude in decimal degrees}

\item{n}{Amount of station to be reviewed}

\item{long}{Longitude in decimal degrees (in NorthAmerica is negative)}
}
\description{
Find the closest stations n from the HYDAT & USGS database based on latitude and longitude
}
