% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatec.R
\name{nc.open.daily}
\alias{nc.open.daily}
\title{Open one daily NCDF file and report the output based on lat, long, var and start and end date.}
\usage{
nc.open.daily(path = NULL, filename, FUN = mean, var, Latitude, Longitude,
  Level = 1, start = -1, end = -1)
}
\description{
Open one daily NCDF file and report the output based on lat, long, var and start and end date.
}
