% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatec.R
\name{nc.open.monthly}
\alias{nc.open.monthly}
\title{Open one monthly NCDF file and report the output based on lat, long, var and start and end date.}
\usage{
nc.open.monthly(path = NULL, filename, var, Latitude, Longitude, start = -1,
  end = -1)
}
\description{
Open one monthly NCDF file and report the output based on lat, long, var and start and end date.
}
