% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOAA.r
\name{NOAA.capture.GHCND}
\alias{NOAA.capture.GHCND}
\title{Capture the information from NOAA GHCND}
\usage{
NOAA.capture.GHCND(noaakey = "zccPPqgbhBjNoKfRCSRVULogLUIMUMdA")
}
\description{
Capture the information from NOAA GHCND
}
