% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOAA.r
\name{NOAA.prepare}
\alias{NOAA.prepare}
\title{Prepare the information from NOAA}
\usage{
NOAA.prepare(par = "Pp", output = "ECm.st.totp")
}
\arguments{
\item{par}{"Pp" for Precipitation, "Tmean" for Avg. Temp, "Tmax" for Max. Temperature, "Tmin" for Min. Temperature}
}
\description{
Prepare the information from NOAA.
}
