% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliar Tools.R
\name{Pair.Expand}
\alias{Pair.Expand}
\title{Function prepare a dataframe to be used in a x,y facet as the functions pairs or hydropairs
Function prepare a dataframe to be used in a x,y facet as the functions pairs or hydropairs}
\usage{
Pair.Expand(data)
}
\description{
Function prepare a dataframe to be used in a x,y facet as the functions pairs or hydropairs
Function prepare a dataframe to be used in a x,y facet as the functions pairs or hydropairs
}
