% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Frequency Analysis.R
\name{PMP}
\alias{PMP}
\title{Estimate the PMP based on Herschfield Methodology}
\usage{
PMP(vector, input_hrs = 24, area_km = 100, number_of_obs_unit = 1)
}
\arguments{
\item{vector}{Vector}

\item{area_km}{Watershed area. The methodology shows a graph for that information; however, it does not adjust the PMP for that correction}

\item{input_hr}{PMP to be defined 24 hrs, 6 hrs or 1 hr/}

\item{number_of_obs_units}{How many observation units are implemented. ie Daily info for 24 hrs PMP is 1.}
}
\description{
PMP Estimation based on Hershfield methodology from Daily values
}
