% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Snow_model.r
\name{PotentialSolar}
\alias{PotentialSolar}
\title{Potential Solar based on EcoHydRology corrected for higher latitudes}
\usage{
PotentialSolar(lat, Jday)
}
\description{
Potential Solar based on EcoHydRology corrected for higher latitudes
}
