% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evap Corrections.R
\name{Radiation}
\alias{Radiation}
\title{Radiation for Morton's methodologies for High latitudes}
\usage{
Radiation(data, constants, ts = "monthly", solar = "sunshine hours",
  Tdew = T, alpha = NULL)
}
\description{
Radiation for Morton's methodologies for High latitudes
}
