% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reanalisys.R
\name{Reanalysis.out}
\alias{Reanalysis.out}
\title{Reanalisys TS compilation}
\usage{
Reanalysis.out(Reanalysis.table, Baseline.start = 1975, Baseline.end = 2005,
  Analysis.start = 2070, Analysis.end = 2099, legend = "down")
}
\arguments{
\item{annual.ts}{Annual time series}

\item{parameter}{Parameter based on EC tables}
}
\value{
return a figure an analysis of the results in percentage
}
\description{
Prepare a Mann-Kendall analysis based on time series.
}
