% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliar Tools.R
\name{RH2DT}
\alias{RH2DT}
\title{Estimates dew point temperature}
\usage{
RH2DT(Temp, RH)
}
\arguments{
\item{Temp}{Value of the air temperature in celcius degree}

\item{RH}{Relative humidity}
}
\value{
Return the relative humidity
}
\description{
Estimates dew point temperature
}
