% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Frequency Analysis.R
\name{rmode}
\alias{rmode}
\title{Statistical mode}
\usage{
rmode(x)
}
\arguments{
\item{x}{Vector to estimate the statistical mode}
}
\value{
Return the statistical mode
}
\description{
Statistical mode - Used to estimate the best regression
}
