% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Snow_model.r
\name{SnowMelt}
\alias{SnowMelt}
\title{SnowMelt model from EcoHydRology corrected for higher latitudes}
\usage{
SnowMelt(Date, precip_mm, Tmax_C, Tmin_C, lat_deg, slope = 0, aspect = 0,
  tempHt = 1, windHt = 2, groundAlbedo = 0.25, SurfEmissiv = 0.95,
  windSp = 2, forest = 0, startingSnowDepth_m = 0,
  startingSnowDensity_kg_m3 = 450)
}
\description{
SnowMelt model from EcoHydRology corrected for higher latitudes
}
