% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Snow_model.r
\name{Solar}
\alias{Solar}
\title{Solar radiation at ground surface from EcoHydRology corrected for higher latitudes}
\usage{
Solar(lat, Jday, Tx, Tn, albedo = 0.2, forest = 0, slope = 0,
  aspect = 0, units = "kJm2d", latUnits = "unknown", printWarn = TRUE)
}
\description{
Solar radiation at ground surface from EcoHydRology corrected for higher latitudes
}
