% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliar Tools.R
\name{stat_smooth_func}
\alias{stat_smooth_func}
\title{Function to present the linear relationship expression in a ggplot graph
Function to present the linear relationship expression in a ggplot graph}
\usage{
stat_smooth_func(mapping = NULL, data = NULL, geom = "smooth",
  position = "identity", ..., method = "auto", formula = y ~ x,
  se = TRUE, n = 80, span = 0.75, fullrange = FALSE, level = 0.95,
  method.args = list(), na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, xpos = NULL, ypos = NULL)
}
\description{
Function to present the linear relationship expression in a ggplot graph
Function to present the linear relationship expression in a ggplot graph
}
