% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECmeteo.R
\name{station.db}
\alias{station.db}
\title{Index table for EC Meteorology}
\format{Index table for EC Meteorology}
\usage{
data(station.db)
}
\description{
Index table for EC Meteorology
}
\references{
EC
}
\author{
Victor Munoz
}
