% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECflow.R
\name{station.db.USGS}
\alias{station.db.USGS}
\title{Index table for EC Meteorology}
\format{Index table for USGS Meteorology}
\usage{
data(station.db.USGS)
}
\description{
Index table for EC Meteorology
}
\references{
USGS
}
\author{
Victor Munoz
}
