% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliar Tools.R
\docType{data}
\name{StatSmoothFunc}
\alias{StatSmoothFunc}
\title{Function to present the linear relationship expression in a ggplot graph (aux)
Function to present the linear relationship expression in a ggplot graph (aux)}
\format{An object of class \code{StatSmooth} (inherits from \code{Stat}, \code{ggproto}) of length 4.}
\usage{
StatSmoothFunc
}
\description{
Function to present the linear relationship expression in a ggplot graph (aux)
Function to present the linear relationship expression in a ggplot graph (aux)
}
\keyword{datasets}
