% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECmeteo.R
\name{UC.dly}
\alias{UC.dly}
\title{Return a list of dataframe with the update undercatchment values for a meteorological parameter (EC) - Daily}
\usage{
UC.dly(var = "Precipitation")
}
\arguments{
\item{var}{Meteorological parameter for undercatchment. Value between: Precipitation (Precipitation), Snow (Snow), Rain (Rain), Min Temperature (Min.Temp), Mean Temperature (Mean.Temp) and Maximum Temperature (Max.Temp)}
}
\description{
Return a list of dataframe with the update undercatchment values for a meteorological parameter (EC) - Daily
}
