% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECmeteo.R
\name{UC}
\alias{UC}
\title{Return a list of dataframe with the update undercatchment values for a meteorological parameter (EC) - Monthly}
\usage{
UC(var = "Precipitation")
}
\arguments{
\item{var}{Meteorological parameter for undercatchment. Value between: Precipitation (Precipitation), Snow (Snow), Rain (Rain), Wind Speed (Wind), Pressure (Pressure), Sea Pressure (Sea.Pressure), Min Temperature (Min.Temp), Mean Temperature (Mean.Temp) and Maximum Temperature (Max.Temp)}
}
\description{
Return a list of dataframe with the update undercatchment values for a meteorological parameter (EC) - Monthly
}
