% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatec.R
\name{var.text}
\alias{var.text}
\title{Prepare the web link to call for the variables associated to each model and assessment report.}
\usage{
var.text(row = 1)
}
\description{
Prepare the web link to call for the variables associated to each model and assessment report.
}
