% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro.R
\name{water_balance}
\alias{water_balance}
\title{Annual water balance variables for one site}
\usage{
water_balance(latitude, ppt, tmean, tmax, tmin, annual = TRUE)
}
\arguments{
\item{latitude}{degrees}

\item{ppt}{vector of length 12, mm}

\item{tmean}{vector of length 12, degrees C}

\item{tmax}{vector of length 12, degrees C}

\item{tmin}{vector of length 12, degrees C}

\item{annual}{should annual sums be returned? (default TRUE; if FALSE monthly values for each of the 5 variables will be returned)}
}
\value{
A named vector of hydrologic variables: PPT, PET, AET, CWD, RAR
}
\description{
Annual water balance variables for one site
}
