% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Watershed tools.R
\name{watershed.poly}
\alias{watershed.poly}
\title{Watershed Delineation with Taudem}
\usage{
watershed.poly(Longitude, Latitude, Area, filename, highdef = F,
  AOI.length = 0.05)
}
\description{
Watershed Delineation with Taudem
}
