% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECmeteo.R
\name{wind.daily2monthly}
\alias{wind.daily2monthly}
\title{Perform a transformation from daily to montly wind speed and wind direction}
\usage{
wind.daily2monthly(ws.ts, wd.deg.ts, max.miss = 5)
}
\arguments{
\item{ws.ts}{Wind Speed Time series in zoo format, wd.deg.ts Wind direction time series in degrees in zoo format, max.miss Maximum of missing days to be considered a complete month.}
}
\description{
Perform a transformation from daily to montly wind speed and wind direction
}
