% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WindRose.R
\name{wind.rose}
\alias{wind.rose}
\title{Wind rose}
\usage{
wind.rose(data, spd, dir, spdres = 2, dirres = 30, spdmin = 2,
  spdmax = 20, spdseq = NULL, palette = "YlGnBu", countmax = NA,
  debug = 0)
}
\description{
Wind rose
}
