% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliar Tools.R, R/climatec.R
\name{zoo2matrix}
\alias{zoo2matrix}
\alias{zoo2matrix}
\title{Transform a zoo time series in a table, available for un-even months}
\usage{
zoo2matrix(ts)

zoo2matrix(ts)
}
\description{
Transform a zoo time series in a table, available for un-even months

Transform a zoo time series in a table
}
