% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mann-kendall analysis.R
\name{zyp.zhang2}
\alias{zyp.zhang2}
\title{Sen's slope algorith based on Zhang}
\usage{
zyp.zhang2(y, x = 1:length(y), conf.intervals = TRUE,
  preserve.range.for.sig.test = TRUE)
}
\description{
Sen's slope algorith based on Zhang
}
